requirements.txt
# Core web framework and ASGI server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database and ORM
sqlalchemy>=2.0.0
pymssql>=2.2.0            # SQL Server driver (no ODBC required)

# Scientific computing and data processing
numpy>=1.24.0
scipy>=1.11.0

# Configuration and environment
python-dotenv>=1.0.0
pytz>=2023.3

# HTTP client for external APIs
requests>=2.31.0

# Azure SDK components (optional - for management features)
azure-identity>=1.15.0
azure-mgmt-web>=7.1.0
azure-mgmt-sql>=3.0.0,<4.0.0

# Video downloading (YouTube-specific yt-dlp removed)

# Note: pyodbc is no longer required - using pymssql for direct SQL Server connections
# Note: SQLite support is built into Python - no additional dependencies needed
